RSP-1701: move stories to TooltIpTrigger
